Fix: Rename fixture namespaces
