+++ test/Fixture/Definition/Definitions/ImplementsDefinitionButIsAbstract/RepositoryDefinition.php
@@ -11,23 +11,15 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\ThrowsExceptionDuringConstruction;
+namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\ImplementsDefinitionButIsAbstract;
 
 use Ergebnis\FactoryBot\Definition\Definition;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
 
-/**
- * Is not acceptable as it throws an exception during construction.
- */
-final class RepositoryDefinition implements Definition
+abstract class RepositoryDefinition implements Definition
 {
-    public function __construct()
-    {
-        throw new \RuntimeException();
-    }
-
-    public function accept(FixtureFactory $factory): void
+    final public function accept(FixtureFactory $factory): void
     {
         $factory->defineEntity(Fixture\FixtureFactory\Entity\Repository::class);
     }
